+++ lib/pghero.rb
@@ -1,9 +1,6 @@
 # dependencies
 require "active_support"
 
-# stdlib
-require "forwardable"
-
 # methods
 require_relative "pghero/methods/basic"
 require_relative "pghero/methods/connections"
@@ -54,20 +51,6 @@ module PgHero
   self.filter_data = ENV["PGHERO_FILTER_DATA"].to_s.size > 0
 
   class << self
-    extend Forwardable
-    def_delegators :primary_database, :aws_access_key_id, :analyze, :analyze_tables, :autoindex, :autovacuum_danger,
-      :best_index, :blocked_queries, :connections, :connection_sources, :connection_states, :connection_stats,
-      :cpu_usage, :create_user, :database_size, :aws_db_instance_identifier, :disable_query_stats, :drop_user,
-      :duplicate_indexes, :enable_query_stats, :explain, :historical_query_stats_enabled?, :index_caching,
-      :index_hit_rate, :index_usage, :indexes, :invalid_constraints, :invalid_indexes, :kill, :kill_all, :kill_long_running_queries,
-      :last_stats_reset_time, :long_running_queries, :maintenance_info, :missing_indexes, :query_stats,
-      :query_stats_available?, :query_stats_enabled?, :query_stats_extension_enabled?, :query_stats_readable?,
-      :rds_stats, :read_iops_stats, :aws_region, :relation_sizes, :replica?, :replication_lag, :replication_lag_stats,
-      :reset_query_stats, :reset_stats, :running_queries, :aws_secret_access_key, :sequence_danger, :sequences, :settings,
-      :slow_queries, :space_growth, :ssl_used?, :suggested_indexes, :suggested_indexes_by_query,
-      :suggested_indexes_enabled?, :system_stats_enabled?, :table_caching, :table_hit_rate, :table_stats,
-      :total_connections, :transaction_id_danger, :unused_indexes, :unused_tables, :write_iops_stats
-
     def time_zone=(time_zone)
       @time_zone = time_zone.is_a?(ActiveSupport::TimeZone) ? time_zone : ActiveSupport::TimeZone[time_zone.to_s]
     end