+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.0.0 (unreleased)
 
 - Fixed error with unlogged sequences on replicas
+- Removed `create_user` and `drop_user` methods
 - Dropped support for Postgres < 14
 - Dropped support for pg_query < 6
 - Dropped support for Sprockets < 4